Multithreading async support
Async may not always use every core, but when enough are created it will. In this case, enough threads are almost guaranteed to be created unless your images is only 50 pixels high.

Currently render2 uses up almost all of the entire CPU's resources. It would be divide the work into as many chunks as there are cores and then distribute those batches. It will be slower though.